90: NPE on SqlJetDB.close http://code.google.com/p/sqljet/issues/detail?id=90
git-svn-id: http://svn.svnkit.com/repos/sqljet@507 06c6b4df-3a75-4b2a-9d74-3e57e8079485

+++ trunk/sqljet-test/src/org/tmatesoft/sqljet/core/AbstractNewDbTest.java
@@ -49,9 +49,13 @@ abstract public class AbstractNewDbTest extends SqlJetAbstractLoggedTest {
     @After
     public void tearDown() throws Exception {
         try {
-            db.close();
+            if (db != null) {
+                db.close();
+            }
         } finally {
-            file.delete();
+            if (file != null) {
+                file.delete();
+            }
         }
     }
